moved methods in wiring class to app class
